requirements.txt
tensorflow==2.19.0
keras-facenet==0.3.2
mtcnn==1.0.0
opencv-python==4.12.0.88
scikit-learn==1.6.1
numpy==2.0.2
pickle-mixin
pandas
flask
Pillow